+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- hankaku to zenkaku
+
 ## [0.0.3] - 2021-03-23
 ### Added
 - show current version